pyproject.toml
[project]
name = "mcp-starter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.12.15",
    "beautifulsoup4>=4.13.4",
    "dotenv>=0.9.9",
    "fastmcp>=2.11.2",
    "google-genai>=1.29.0",
    "markdownify>=1.1.0",
    "pillow>=11.3.0",
    "python-dotenv>=1.1.1",
    "readabilipy>=0.3.0",
    "scikit-image>=0.25.2",
    "spotipy>=2.25.1",
    "youtube-transcript-api>=1.2.2",
    "yt-dlp[curl-cffi,default]>=2025.7.21",
]
